test(std/uuid): use one test module per public module
