+++ std/uuid/test.ts
@@ -1,17 +1,13 @@
-#!/usr/bin/env -S deno run
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
+import { assert } from "../testing/asserts.ts";
+import { NIL_UUID, isNil } from "./mod.ts";
 
-// Generic Tests
-import "./tests/isNil.ts";
-
-// V1 Tests
-import "./tests/v1/validate.ts";
-import "./tests/v1/generate.ts";
-
-// V4 Tests
-import "./tests/v4/validate.ts";
-import "./tests/v4/generate.ts";
-
-// V5 Tests
-import "./tests/v5/validate.ts";
-import "./tests/v5/generate.ts";
+Deno.test({
+  name: "[UUID] isNil",
+  fn(): void {
+    const nil = NIL_UUID;
+    const u = "582cbcff-dad6-4f28-888a-e062ae36bafc";
+    assert(isNil(nil));
+    assert(!isNil(u));
+  },
+});
