+++ std/uuid/v4_test.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
-import { assert, assertEquals } from "../../../testing/asserts.ts";
-import { generate, validate } from "../../v4.ts";
+import { assert, assertEquals } from "../testing/asserts.ts";
+import { generate, validate } from "./v4.ts";
 
 Deno.test({
   name: "[UUID] test_uuid_v4",
@@ -18,5 +18,18 @@ Deno.test({
       const u = generate() as string;
       assert(validate(u), `${u} is not a valid uuid v4`);
     }
+  },
+});
+
+Deno.test({
+  name: "[UUID] is_valid_uuid_v4",
+  fn(): void {
+    const u = generate();
+    const t = "84fb7824-b951-490e-8afd-0c13228a8282";
+    const n = "84fb7824-b951-490g-8afd-0c13228a8282";
+
+    assert(validate(u), `generated ${u} should be valid`);
+    assert(validate(t), `${t} should be valid`);
+    assert(!validate(n), `${n} should not be valid`);
   },
 });
